+++ src/app/modules/languages/index.js
@@ -10,8 +10,14 @@ app.config(EnglishProvider);
 import ChineseProvider from './cn';
 app.config(ChineseProvider);
 
+import ItalianProvider from './it';
+app.config(ItalianProvider);
+
 import PolishProvider from './pl';
 app.config(PolishProvider);
+
+import PortugueseBRProvider from './ptbr';
+app.config(PortugueseBRProvider);
 
 import GermanProvider from './de';
 app.config(GermanProvider);
@@ -24,6 +30,12 @@ app.config(RussianProvider);
 
 import DutchProvider from './nl';
 app.config(DutchProvider);
+
+import SpanishProvider from './es';
+app.config(SpanishProvider);
+
+import UkrainianProvider from './uk';
+app.config(UkrainianProvider);
 
 // Comment this while developing to see untranslated strings
 app.config(['$translateProvider', function($translateProvider) {