+++ src/app/modules/languages/jp.js
@@ -27,8 +27,8 @@ function JapaneseProvider($translateProvider) {
       HEADER_NODE_CUSTOM_TOOLTIP: 'あなた自身のNISノードを入力',
       HEADER_PURGE: "アカウントデータの削除(パージ)",
       HEADER_PURGE_MESSAGE: "必ずバックアップを行った上で実行して下さい。OK をクリックすると、ローカルストレージに登録されているすべてのウォレットが削除(パージ)されます。すべてのウォレットをバックアップしていた場合、残高は安全です（回復にはバックアップファイルの再読み込みが必要です）。",
-      HEADER_OFFLINE_TX: 'Prepare an offline transaction',
-      HEADER_RELEASE_TX: 'Release a transaction',
+      HEADER_OFFLINE_TX: 'オフライントランザクションの準備',
+      HEADER_RELEASE_TX: 'トランザクションのリリース',
 
       // FOOTER COMPONENT
       FOOTER_POWERED_BY_1: 'Powered by',
@@ -59,7 +59,8 @@ function JapaneseProvider($translateProvider) {
       DASHBOARD_NOTICE_3: 'NEM チームは、メインネットを利用する前に、最初にテストネットにて NanoWallet の操作を行い、利用方法を習得する事をお勧めします。英文のガイドブックは<a href="https://blog.nem.io/nanowallet-tutorial/" target="_blank">こちら</a>にあります。利用者は自身の資金を管理し、秘密鍵をバックアップしてください。NEM チームは、メインネット上でこのアプリケーションを利用することに起因する、資金の消失について責任を負いません。',
       DASHBOARD_NOTICE_4: '紙に秘密鍵を書き留めて、安全な場所に保存することをお勧めします。また、あなたのアカウントをAndroid 及び iOS アプリにインポートして、バックアップする事もできます。',
       DASHBOARD_NOTICE_5: 'Thanks for your participation!', //すまねぇ、ここどう訳すか迷ってそのままｗ
-      
+      DASHBOARD_MORE_TXES: 'もっとトランザクションを見る',
+
       // GENERAL
       GENERAL_BLOCK: "ブロック",
       GENERAL_BLOCKS: 'ブロック',
@@ -161,8 +162,10 @@ function JapaneseProvider($translateProvider) {
       GENERAL_ENCRYPTED: '暗号化',
       GENERAL_UNENCRYPTED: '平文',
       GENERAL_HEXADECIMAL: '16進数',
-      GENERAL_SELECT_ACCOUNT: 'Select account',
-      GENERAL_INVOICE: 'Invoice',
+      GENERAL_SELECT_ACCOUNT: 'アカウントの選択',
+      GENERAL_INVOICE: '請求書',
+      GENERAL_SIGNATURE: '署名',
+      GENERAL_VERIFY: '検証',
 
       // HOME MODULE
       HOME_UNSUPPORTED_BROWSER: '申し訳ありません。このブラウザでは安全に NanoWallet を使用する事ができません。',
@@ -370,6 +373,8 @@ function JapaneseProvider($translateProvider) {
       ACCOUNT_CUSTOM_NODE: "カスタムノードを使う",
       ACCOUNT_NODE_FROM_LIST: "一覧からノードを使う",
       ACCOUNT_DELEGATED_PRIVATE_KEY: "委任秘密鍵",
+      ACCOUNT_NO_PUBLIC_KEY: '公開鍵を取得するためにはトランザクションを発行してください。',
+      ACCOUNT_SHOW_ON_TREZOR_BTN: 'TREZORで見る',
 
       // PORTAL MODULE
       PORTAL_TITLE: "各種機能",
@@ -380,9 +385,9 @@ function JapaneseProvider($translateProvider) {
       PORTAL_MULTISIG_BTN_3: 'マルチシグトランザクションの署名',
       PORTAL_HARVESTING_TITLE: "デリゲートハーベスティング（委任）",
       PORTAL_HARVESTING_TEXT: "デリゲートハーベスティングは、アカウントが起動していない時間でもリモート・ノードを介して「マイニング（ハーベスト）」を可能にする機能です。",
-      PORTAL_CHANGELLY_TITLE: "Changelly （XEM 等の両替サイト）",
-      PORTAL_CHANGELLY_TEXT: "最適なレートで XEM を交換できる Changelly ウィジェットを使用できます。",
-      PORTAL_CHANGELLY_BTN: "XEM を購入",
+      PORTAL_EXCHANGE_TITLE: 'インスタントエクスチェンジ',
+      PORTAL_EXCHANGE_TEXT: 'ChangellyやShapeShiftウィジェットを使って、最適レートでXEMを購入しましょう！',
+      PORTAL_EXCHANGE_BTN: 'XEMを購入',
       PORTAL_NS_TITLE: "ネームスペースとサブドメイン",
       PORTAL_NS_TEXT: "ネームスペースはドメイン名のようなものです。ネームスペース名はユニーク(他に同一の名称無し)であり、その配下にサブネームスペースやモザイク （アセット） を発行する事ができます。",
       PORTAL_NS_BTN: "ネームスペースを作成",
@@ -396,7 +401,8 @@ function JapaneseProvider($translateProvider) {
       PORTAL_APOSTILLE_BTN_2: "監査（公証の確認）",
       PORTAL_ADDRESS_BOOK_TEXT: 'アドレスにラベルを付けることによって、 連絡先を簡単に管理する事ができます。',
       PORTAL_ADDRESS_BOOK_BTN: 'アドレス帳の管理',
-      PORTAL_INVOICE_TEXT: 'Create an invoice to share via QR code',
+      PORTAL_INVOICE_TEXT: '共有するためのQRコード請求書を作成します',
+      PORTAL_SIGNED_MSG_TEXT: 'トランザクションなしにアカウントの所有認証をするための署名済みメッセージの作成と検証を行います。',
 
       // ADDRESS BOOK MODULE
       ADDRESS_BOOK_TITLE: 'アドレス帳',
@@ -441,6 +447,11 @@ function JapaneseProvider($translateProvider) {
       // ACCOUNT EXPLORER
       ACCOUNTS_EXPLORER_TITLE: 'エクスプローラ - アカウント',
       ACCOUNTS_EXPLORER_SEARCH: '検索',
+
+      // TRANSACTIONS EXPLORER
+      EXPLORER_TRANSACTIONS_TITLE: 'あなたのトランザクション履歴',
+      EXPLORER_TRANSACTIONS_LOAD_MORE: 'もっと読み込む',
+
 
       // APOSTILLE HISTORY MODULE
       APOSTILLE_HISTORY_TITLE: "アポスティーユの履歴",
@@ -603,7 +614,8 @@ function JapaneseProvider($translateProvider) {
       ALERT_ADDRESS_BOOK_FILE_SUCCESS: 'アドレス帳のインポートに成功しました!',
       ALERT_VOTING_SUCCESS: '投票の送信に成功しました。',
       ALERT_POLL_CREATION_SUCCESS: '投票の作成に成功しました。',
-      ALERT_COPY_SIGNED_TX_SUCCESS: 'Signed transaction copied!',
+      ALERT_COPY_SIGNED_TX_SUCCESS: '署名済みトランザクションをコピーしました！',
+      ALERT_COPY_SIGNED_MSG_SUCCESS: '署名済みメッセージをコピーしました！',
 
       // CONVERT ACCOUNT TO MULTISIG
       AGGREGATE_MODIFICATION_TITLE: "マルチシグアカウントへ変換",
@@ -670,7 +682,7 @@ function JapaneseProvider($translateProvider) {
       SIGNUP_BRAIN_WALLET_WARNING: "ブレインウォレットの<a href=\"https://en.bitcoin.it/wiki/Brainwallet\" rel=\"nofollow\" target=\"_blank\">危険性</a>について留意してください。ブレインウォレットはハッシュ化されたパスフレーズのみを利用し何度も使用されます。したがって、少なくとも40文字以上の「安全な」パスフレーズを選んでください。<a href=\"https://xkcd.com/936/\" rel=\"nofollow\" target=\"_blank\">XKCD #936</a>",
       SIGNUP_PRIVATE_KEY_WALLET_WARNING: "プライベートキーウォレットは、インポートされた秘密鍵を暗号化するためにパスワードのみを使用します。したがって、「安全な」パスワードを利用する事が重要です。 ",
       SIGNUP_CREATE_START_WARNING: '各ステップは慎重に行ってください！',
-      SIGNUP_CREATE_START_CONNECTION_WARNING: 'It is recommended to disconnect from internet while creating your wallet and backing up your data.',
+      SIGNUP_CREATE_START_CONNECTION_WARNING: 'ウォレットの作成とデータのバックアップ中にはインターネットから切断することを推奨します。',
       SIGNUP_CREATE_READY_BTN: '準備できました',
       SIGNUP_CREATE_ENTER_NAME: 'ウォレット名を入力してください',
       SIGNUP_CREATE_ENTER_PASSWORD: 'パスワードを入力してください',
@@ -692,15 +704,16 @@ function JapaneseProvider($translateProvider) {
       SIGNUP_COMMON_WALLET_WARNING_BTN_1: 'ウォレットファイルの表示',
       SIGNUP_COMMON_WALLET_WARNING_BTN_1_INFO: 'ウォレットファイルを作成するには、空のテキストファイルを作成し、上記のBase64で作成されたキーを記入します。ファイルを<b><i>yourWalletName.wlt</i></b>という名前で保存し、インポートすることができます。',
       SIGNUP_COMMON_WALLET_WARNING_BTN_2: 'プライベートキーの表示',
-      SIGNUP_COMMON_WALLET_WARNING_BTN_2_INFO: '秘密鍵をバックアップするには、テキストファイルに保存して印刷するか、安全な場所に書き込んでください。 秘密鍵はオフラインで保存することをお勧めします。',
+      SIGNUP_COMMON_WALLET_WARNING_BTN_2_INFO: '秘密鍵はアカウントの全ての権限を持つので必ず秘密にしてください。<b>絶対に誰にも教えてはいけません</b>。秘密鍵は安全でオフラインな場所に保管することを<b>強く推奨します</b>。',
       SIGNUP_COMMON_WALLET_WARNING_FOOTER: '以下をクリックすると、上記全ての警告を読み理解し同意したことになります。',
       SIGNUP_COMMON_WALLET_WARNING_CONFIRM_1: 'ウォレットファイルをバックアップしました',
-      SIGNUP_COMMON_WALLET_WARNING_CONFIRM_2: '秘密鍵とパスワードバックアップしました',
+      SIGNUP_COMMON_WALLET_WARNING_CONFIRM_2: '秘密鍵とパスワードをバックアップしました',
       SIGNUP_COMMON_WALLET_WARNING_CONFIRM_3: '私は全ての内容を理解し同意しました',  //このあたりは、英文より厳しく書いてます。
       SIGNUP_COMMON_WALLET_WARNING_UNDERSTOOD: '理解し同意しました',
       SIGNUP_COMMON_WALLET_WARNING_DOWNLOAD: 'ウォレットをダウンロードする',
       SIGNUP_ESTIMATED_PASSPHRASE_STRENGTH: '推定パスフレーズ強度',
       SIGNUP_ESTIMATED_GUESS_TIMES: '推定解析時間',
+      SIGNUP_CONGRATS_MSG: '<b>おめでとう！</b> あなたの NEM アドレスは',
 
       // FAQ MODULE
       FAQ_TITLE: "よくある質問",
@@ -724,18 +737,18 @@ function JapaneseProvider($translateProvider) {
       FAQ_ANSWER_8: 'この場合、"サービス"へ移動し、"マルチシグおよびマルチユーザーアカウント"の項目から"マルチシグトランザクションの署名"を選択してください。',
       FAQ_QUESTION_9: '最適なセキュリティは？',
       FAQ_ANSWER_9: '特に推奨するのは秘密鍵を紙に保存することです。<br>プリントアウトしてどこか安全な所で保管してください。<br><br>ウォレットファイルについても同様に、複数のコピーをUSBメモリなどのオフラインな場所に保存してください。<br>パスワードは一意で複雑にしてください。そのため、まずは書き留めておいてください。<br><br>ウォレットを確認する、もしくは操作をしたいときは:<br> - USBメモリを指します<br> - NanoWalletにインポートします<br> - USBメモリを抜きます<br><br> ウォレットのコピーは必要な時にブラウザのローカルストレージに保存されます。<br>作業が終わったら、ログアウトし、フッターの右端にある削除ボタンでローカルストレージ内のウォレットを削除してください。',
-      FAQ_QUESTION_10: 'Where to find information about my account (address, etc) ?',
-      FAQ_ANSWER_10: 'If you look at the top navigation bar, you will see an <b><i>"Account"</b></i> button between <b><i>"Node"</b></i> and <b><i>"Language"</b></i>. There you can find your address, public key, vested balance and other important data.',
-      FAQ_QUESTION_11: 'I have deposited XEM to an exchange but nothing has been credited ?',
-      FAQ_ANSWER_11_1: 'First you must check that the hash of your transaction is pointing to an existing transaction on the <a target="_blank" href="http://chain.nem.ninja">explorer</a> (please, note that the explorer is a few blocks behind).',
-      FAQ_ANSWER_11_2: 'Most exchanges are asking for an identification message to credit your deposit. Make sure that you have followed carefully the exchange instructions and added a message that is NOT encrypted. ',
-      FAQ_ANSWER_11_3: 'Even if you have added a message, it can happen that exchanges may not process your deposit, because of issues on their side.',
-      FAQ_ANSWER_11_4: 'You must contact the exchange support, explain the situation and provide them the hash of the transaction.',
-      FAQ_QUESTION_12: 'How to check if I am on a fork ?',
-      FAQ_ANSWER_12_1: 'Click on <b><i>"Node"</b></i>, in the top navigation bar, to open the node panel.',
-      FAQ_ANSWER_12_2: 'Look at the chain height and compare it to the height shown <a target="_blank" href="http://bigalice3.nem.ninja:7890/chain/height">here</a>.',
-      FAQ_ANSWER_12_3: 'If different of more than 5 blocks then you are probably on a fork.',
-      FAQ_ANSWER_12_4: 'To fix, just choose another node from the dropdown in the node panel, it will restore your account at it\'s latest state on the real network.',
+      FAQ_QUESTION_10: 'どこに自分のアカウントの情報(アドレスなど)がありますか？',
+      FAQ_ANSWER_10: '上部のナビゲーションバーにある<b><i>"ノード"</b></i>と<b><i>"言語"</b></i>の間に<b><i>"アカウント"</b></i>ボタンがあります。ここでアドレスや公開鍵、既得バランスなどの重要な情報を得られます。',
+      FAQ_QUESTION_11: '交換所にXEMを入金しましたが、残高に反映されません。',
+      FAQ_ANSWER_11_1: '最初にあなたのトランザクションを<a target="_blank" href="http://chain.nem.ninja">エクスプローラー</a>上に表示されるか確認してください。(エクスプローラーには数ブロック程度の遅延があります)',
+      FAQ_ANSWER_11_2: 'ほとんどの交換所はあなたの入金を識別するためにメッセージを要求します。各交換所においての手順をよく理解し、暗号化していないメッセージを添付してください。',
+      FAQ_ANSWER_11_3: '交換所の問題によって、メッセージを添付しているにも関わらずあなたの入金が処理されない場合もあります。',
+      FAQ_ANSWER_11_4: '交換所のサポートへトランザクションのハッシュを提供して状態を説明してください。',
+      FAQ_QUESTION_12: 'フォーク上にいるかを確認するためには？',
+      FAQ_ANSWER_12_1: '上部のナビゲーションバーから<b><i>"ノード"</b></i>をクリックし、ノードパネルを開きます。',
+      FAQ_ANSWER_12_2: 'ブロック高について<a target="_blank" href="http://bigalice3.nem.ninja:7890/chain/height">ここ</a>で表示できる高さと比較してみてください。',
+      FAQ_ANSWER_12_3: 'もし5ブロック以上差がある場合はフォークしているかもしれません。',
+      FAQ_ANSWER_12_4: '補正するためにはノードパネルのドロップダウンリストから他のノードを選択してください。ネットワーク上の最新の情報でアカウントが再読込されます。',
 
       // FORM RELATED
       FORM_PASSWORD_FIELD_PLACEHOLDER: "ウォレットのパスワード、もしくはパスフレーズを入力してください。",
@@ -776,26 +789,82 @@ function JapaneseProvider($translateProvider) {
       FORM_SELECT_MULTISIG: 'Select a multisignature account',
       FORM_SELECT_CONTACT: 'Select a contact',
 
+      // TREZOR RELATED
+      TREZOR_TITLE: 'TREZOR',
+      TREZOR_TEXT: 'ハードウェアウォレットのTREZORはXEM、モザイク、マルチシグアカウントを保護できます。',
+      TREZOR_BUTTON: 'TREZORでログイン',
+
       // CREATE OFFLINE TRANSACTION MODULE
-      OFFLINE_TX_TITLE: 'Prepare an offline transaction',
-      OFFLINE_TX_NO_WALLET: 'Please import a wallet from login module to see the form.',
-      OFFLINE_TX_INFO_1: 'Make sure to be disconnected of internet when importing your wallet and creating the transaction!',
-      OFFLINE_TX_INFO_2: 'Only simple transactions can be created because a connection is needed to fetch mosaics and multisig information from NEM nodes.',
-      OFFLINE_TX_INFO_3: 'After clicking the "create" button in the left panel, you will find the signed transaction below.',
-      OFFLINE_TX_INFO_4: 'A signed transaction is immutable and will be effective only if released to the network before the default deadline of 24 hours.',
-      OFFLINE_TX_SIGNED: 'Signed transaction',
-      OFFLINE_TX_RELEASE: 'Open in release module',
+      OFFLINE_TX_TITLE: 'オフライントランザクションの準備',
+      OFFLINE_TX_NO_WALLET: 'フォームを表示するためにログイン画面からウォレットをインポートしてください。',
+      OFFLINE_TX_INFO_1: '安全性を確保するため、コンピュータがインターネットに<b>接続していない</b>ことを確認してください。',
+      OFFLINE_TX_INFO_2: 'NEMノードからモザイクやマルチシグ情報を取得するにインターネット接続が必要なので、単純なトランザクションだけ作成できます。',
+      OFFLINE_TX_INFO_3: '左側のパネルの"作成"ボタンをクリックすると、下部に署名済みトランザクションが表示されます。',
+      OFFLINE_TX_INFO_4: '署名済みトランザクションは変更不可であり、デフォルト期限の24時間以内にリリースされた場合にだけ有効になります。',
+      OFFLINE_TX_SIGNED: '署名済みトランザクション',
+      OFFLINE_TX_RELEASE: 'リリースモジュールで開く',
+      OFFLINE_TX_MODAL_ALERT: 'このコンピュータはインターネットに接続されているようです。安全にオフライントランザクションを作るために<b>インターネットに接続されていないコンピュータ上だけ</b>で行ってください。',
 
       // RELEASE OFFLINE TRANSACTION MODULE
-      RELEASE_OFFLINE_TX_TITLE: 'Release a transaction to the network',
-      RELEASE_OFFLINE_TX_PARAMETERS: 'Transaction parameters',
-      RELEASE_OFFLINE_TX_INFO_1: 'To release a transaction you must be connected to internet.',
-      RELEASE_OFFLINE_TX_INFO_2: 'Make sure that you have selected the right network and a working node or it will be rejected.',
-      RELEASE_OFFLINE_TX_INFO_3: 'It is not possible to send a signed transaction twice. One signed transaction will always generate the same hash and two transactions cannot have the same hash.',
-      RELEASE_OFFLINE_TX_INFO_4: 'You can release a signed transaction from any computer safely.',
+      RELEASE_OFFLINE_TX_TITLE: 'トランザクションをネットワークにリリース',
+      RELEASE_OFFLINE_TX_PARAMETERS: 'トランザクションのパラメータ',
+      RELEASE_OFFLINE_TX_INFO_1: 'トランザクションをリリースするにはインターネットへの接続が必要です。',
+      RELEASE_OFFLINE_TX_INFO_2: '対応するネットワークの稼働中ノードを選択していることを確認してください。間違っている場合は拒否されます。',
+      RELEASE_OFFLINE_TX_INFO_3: '署名済みトランザクションを2度送信することはできません。トランザクションは常に同じハッシュを生成し、2つのトランザクションは同じハッシュを持つことができません。',
+      RELEASE_OFFLINE_TX_INFO_4: '署名済みトランザクションはいかなるコンピュータからでも安全にリリースできます。',
 
       // INVOICE MODULE
-      CREATE_INVOICE_TITLE: 'Create an invoice'
+      CREATE_INVOICE_TITLE: '請求書を作成',
+
+      // CREATE SIGNED MESSAGE MODULE
+      CREATE_SIGNED_MSG_TITLE: "署名済みメッセージの作成",
+      CREATE_SIGNED_MSG_NAME: '署名済みメッセージ',
+      CREATE_SIGNED_MSG_SIGN: 'メッセージに署名',
+
+      // VERIFY SIGNED MESSAGE MODULE
+      VERIFY_SIGNED_MSG_TITLE: "署名済みメッセージの検証",
+      VERIFY_SIGNED_MSG_SIG_VALID: '署名は有効です',
+      VERIFY_SIGNED_MSG_SIG_NOT_VALID: '署名は<b>不正</b>です',
+      VERIFY_SIGNED_MSG_NOT_VALID: '署名済みメッセージが不正です！',
+
+      // DNS MODULE
+      DNS_SHORT_DESC: "namespace.nemドメイン用の分散ドメインネームシステム(DNS)サービス",
+      DNS_CONFIGURE: "DNSの設定",
+      DNS_SEARCH_DNS: "DNSの検索",
+      DNS_GO_SEARCH: "検索",
+      DNS_GET_INFO: "情報取得",
+      DNS_SEARCH_HEADLINE: "DNSの検索",
+      DNS_SEARCH_HELP: "DNSのヘルプ",
+      DNS_SEARCH: "検索",
+      DNS_SEARCH_HELP_INFO: "namespace.nemを入力して関連するDNS情報を参照するか、直接ウェブサイトに移動します。",
+      //DNS CONFIGURE
+      DNS_SELECT_NAMESPACE:"ネームスペースの選択",
+      DNS_POINTER_ADDRESS:"ポインタアドレス",
+      DNS_IP_ADDRESS:"IPアドレス",
+      DNS_ORGA:"組織",
+      DNS_COUNTRY:"国籍",
+      DNS_ADDRESS:"住所",
+      DNS_PHONE:"電話番号",
+      DNS_EMAIL:"Eメール",
+      DNS_OTHERINFO:"その他",
+      DNS_JSONMSG:"JSONメッセージ",
+      DNS_CHAR_LEFT:"残り文字数",
+      DNS_HELP1:"NEMネームスペースの所有者はNEM DNSが簡単な数ステップで使用できます。DNSはNEMエコシステムの名前空間プラットフォームと、関連付けられたポインターアドレス(PA)を中心に構築されています。ネームスペースの所有者は、ネームスペースに関連する情報を使用して、JSON形式の暗号化されていないメッセージを含む、ネームスペースポインタアドレスにトランザクションを送信する必要があります。メッセージには、IPアドレス、所有権情報、物理アドレス、連絡先情報などの関連情報を含めることができます。",
+      DNS_HELP2:"このように使います:",
+      DNS_HELP3:"1. ネームスペースを選ぶ",
+      DNS_HELP4:"2. 関連を書き込むかDNS情報を更新する",
+      DNS_HELP5:"3. ポインタアドレスへトランザクションを送る",
+      //DNS ALERTS
+      ALERT_DNS_SUCCESS: "情報が見つかりました",
+      ALERT_DNS_NO_INFO_FOUND: "DNS情報が見つかりませんでした",
+
+      // IMPORT WALLET QRCODE
+      IMPORT_WALLET_QRCODE_TITLE: 'ウォレットQRコードのインポート',
+      IMPORT_WALLET_QRCODE_INFORMATION: 'サポートしているウォレットQRコード: NEM wechat wallet, ios, android.',
+      IMPORT_WALLET_QRCODE_LOAD: 'ウォレットQRコードを読み込む',
+      IMPORT_WALLET_QRCODE_SCAN: 'スキャン',
+      IMPORT_WALLET_QRCODE_STOP_SCAN: 'スキャン停止'
+
     });
 
 }