+++ src/app/utils/languages.js
@@ -2,26 +2,40 @@
  * Available languages
  */
 let data =  [{
-    name: "English",
-    key: "en"
-}, {
     name: "Chinese",
     key: "cn"
 }, {
-    name: "Polish",
-    key: "pl"
-},{
+    name: "English",
+    key: "en"
+}, {
     name: "German",
     key: "de"
-}, {
+}, 
+  {
+    name: "Italian",
+    key: "it"
+},              
+ {
     name: "Japanese",
     key: "jp"
+}, {
+    name: "Nederlands",
+    key: "nl"
+}, {
+    name: "Polish",
+    key: "pl"
+}, {
+    name: "Português (Brasil)",
+    key: "ptbr"
 }, {
     name: "Русский",
     key: "ru"
 }, {
-    name: "Nederlands",
-    key: "nl"
+    name: "Spanish",
+    key: "es"
+}, {
+    name: "Ukrainian",
+    key: "uk"
 }]
 
 module.exports = {